requirements.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
click==6.7
Flask==0.12.2
funcsigs==1.0.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
mock==2.0.0
numpy==1.12.1
olefile==0.44
packaging==16.8
pbr==3.0.1
Pillow==4.1.1
protobuf==3.3.0
pyparsing==2.2.0
requests==2.16.5
six==1.10.0
tensorflow==1.1.0
urllib3==1.21.1
Werkzeug==0.12.2
